+++ src/components/GenePage.css
@@ -19,13 +19,6 @@
     padding-right: 0;
 }
 
-.codeCard{
-    border-radius: 8px;
-    box-shadow: 0px 0px 8px #ccc;
-    padding: 5px;
-    display: inline;
-}
-
 #table_options_outer{
     display: flex;
     justify-content: center;
