+++ src/components/GeneSequenceAnimation.css
@@ -1,0 +1,7 @@
+
+.codeCard{
+    border-radius: 8px;
+    box-shadow: 0px 0px 8px #ccc;
+    padding: 5px;
+    display: inline;
+}
